app/requirements.txt
click==6.7
Flask-SQLAlchemy==2.3
Flask==0.12.2
Flask-WTF==0.14.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-MySQL==1.4.0
mysql-connector-python==8.0.6
PyMySQL==0.8.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1

